Refactoring magna flow handling. (#35)
* Refactoring magna flow handling.

This commit splits out counter and reporter from
specific mpls handling such that they can be used
by other packet generators.

* IP Traffic generator

This commit adds a simple IP traffic generator function. This
supports both IPv4 and IPv6 but not IPv4 in IPv6 and vice versa.
